+++ CHANGELOG.md
@@ -10,6 +10,7 @@
   - Regression fix on onProgress(int,int) documentation
   - Sample application now can be built with LeakCanary, use i.e. `gradle :sample:installWithLeakCanaryDebug` to use it
   - Updated RequestParams documentation on handling arrays, sets and maps, along with new RequestParamsDebug sample
+  - Added BlackholeHttpResponseHandler implementation, which discards all response contents and silents all various log messages (see #416)
 
 ## 1.4.7 (released 9. 5. 2015)